undo riskier changes
